xml: Parsing UDonkey's XML
MainDB.xml included, as it can be annoying to generate without Windows. This
should be removed from git before merging into master.

Course sports display almost-properly (their even type is malformed, it seems),
but do not get scheduled for some reason.

Importing with rexml is slow (as slow as REPY parsing!), we need to find out
if we can either use something faster or show progress properly.

The old parsing/logic combination is really horrible it seems, it must be
cleaned up.